requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0

# Cloud SDKs
boto3==1.34.34
google-cloud-storage==2.14.0
azure-storage-blob==12.19.0

# Queue & Cache
redis==5.0.1
kafka-python==2.0.2

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# HTTP & Networking
httpx==0.26.0
aiohttp==3.9.1
dnspython==2.5.0
requests==2.31.0

# Rate Limiting & Proxy
aiolimiter==1.1.0
python-socks==2.4.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
pyyaml==6.0.1
tldextract==5.1.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
